Moved Schema tests to relevant folder
