+++ test/end-to-end/applications/company/:id/reject.js
